python crud project project/requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
pydantic==2.11.5
python-dotenv==1.1.0
flask==3.0.0